Cleaning up solution
Default nuget build now AnyCPU
Created a separate unit test project
NUnit and NuSpec no longer listed as dependencies on package
