+++ packages.config
@@ -2,6 +2,5 @@
 <packages>
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
   <package id="NuGet.Build.Packaging" version="0.2.0" targetFramework="net45" developmentDependency="true" />
-  <package id="NUnit" version="3.9.0" targetFramework="net45" />
-  <package id="NuSpec" version="3.0.0" targetFramework="net45" />
+  <package id="NuSpec" version="3.0.0" targetFramework="net45" developmentDependency="true" />
 </packages>
