+++ pokitdok-csharp.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.27703.2042
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pokitdok-csharp", "pokitdok-csharp.csproj", "{BE56DFAC-58A2-4B62-AF3D-8CDDA98E6941}"
 EndProject
@@ -9,6 +9,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{488A57
 	ProjectSection(SolutionItems) = preProject
 		.nuget\packages.config = .nuget\packages.config
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pokitdok-csharp.tests", "pokitdok-csharp.tests\pokitdok-csharp.tests.csproj", "{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -26,8 +28,19 @@ Global
 		{BE56DFAC-58A2-4B62-AF3D-8CDDA98E6941}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BE56DFAC-58A2-4B62-AF3D-8CDDA98E6941}.Release|x86.ActiveCfg = Release|x86
 		{BE56DFAC-58A2-4B62-AF3D-8CDDA98E6941}.Release|x86.Build.0 = Release|x86
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Debug|x86.Build.0 = Debug|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Release|x86.ActiveCfg = Release|Any CPU
+		{F719DC09-EE74-4E16-87D3-A3AE67DF9AE1}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C79456EF-5465-48B8-BD59-8629444780B0}
 	EndGlobalSection
 EndGlobal
