+++ pokitdok-csharp.tests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net45" />
+  <package id="NUnit" version="3.9.0" targetFramework="net45" developmentDependency="true" />
+</packages>
